ci: SM-70 add different routing
