+++ src/controllers/user_controller.py
@@ -1,11 +1,25 @@
-from src.controllers.controller import router
+from fastapi import APIRouter, HTTPException
+from src.database.database_utils import get_all
+
+from src.models.user_model import User
 from src.schemas.user_schema import UserSchema
 from src.services.user_service import check_password
 
+router = APIRouter()
 
-@router.post("/login/", response_model=UserSchema)
+
+@router.get("/allUser")
+async def get_all_entries():
+    return get_all(User)
+
+
+@router.post("/login")
 async def post_login(string: UserSchema):
     id = string.id
     password = string.password
     result = check_password(id, password)
-    return f"The password is {result}"
+
+    if result:
+        return {"message": "Login successful"}
+    else:
+        raise HTTPException(status_code=401, detail="Incorrect password or user")
