+++ src/main.py
@@ -1,9 +1,11 @@
 from .database.database_setup import init_db
 from fastapi import FastAPI
-from .controllers.controller import router_func
+from .controllers import password_controller, user_controller
 
 
 app = FastAPI()
 
 init_db()
-router_func(app)
+
+app.include_router(password_controller.router)
+app.include_router(user_controller.router)
